storage_stack/requirements.txt
psycopg2-binary
SQLAlchemy
pymongo
python-dotenv
